requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
click==8.0.3
Django==3.2
django-debug-toolbar==3.2.2
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.0
et-xmlfile==1.1.0
gunicorn==20.1.0
h11==0.12.0
inflection==0.5.1
jsonschema==4.2.1
number-precision==2021.1.30
openpyxl==3.0.9
pendulum==2.1.2
Pillow==8.4.0
pycodestyle==2.8.0
PyJWT==2.3.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pytzdata==2020.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
uvicorn==0.16.0
